Updating framerate management
Now the lag is registered globally (and not only the delta
time to the frame before), which allows to skip frames
or to sleep more precisely, if lags accumulate slowly

Tested on a 120fps 1080p video, it skipped a lot of frames
but the video speed was accurate. On small SD videos, it
was displaying the right framerate, without skip and without
going too fast.