[IMPROVEMENT] FPS increase
Two way used to increase fps:
-use of av resize integrated with "to_image" method. It's
made to resize video stream and is far more optimized than
PIL for this usecase
-use of imagetk paste method. Display frame was creating
a new PhotoImage each call, but a PhotoImage call is just
a imagetk creation of the good dimension/type, then a
paste of the content inside it. Now we create a new PhotoImage
only if the dimension changes.
It also seems that less call to the Tk event loop are made
when we paste new content instead of creating a new PhotoImage